requirements.txt
nltk
openai
pandas
pinecone
pinecone_text
streamlit
google-generativeai
